Cleanup folder structures, remove redundant code
