+++ lib/model/model.go
@@ -147,10 +147,8 @@ func (m *Model) StartDeadlockDetector(timeout time.Duration) {
 	deadlockDetect(m.pmut, timeout)
 }
 
-// StartFolderDefault starts read/write processing on the current model. When in
-// read/write mode the model will attempt to keep in sync with the cluster by
-// pulling needed files from peer devices.
-func (m *Model) StartFolderDefault(folder string) {
+// StartFolder starts processing on the current model.
+func (m *Model) StartFolder(folder string) {
 	m.fmut.Lock()
 	cfg, ok := m.folderCfgs[folder]
 	if !ok {
@@ -161,11 +159,12 @@ func (m *Model) StartFolderDefault(folder string) {
 	if ok {
 		panic("cannot start already running folder " + folder)
 	}
-	p := newDefaultFolder(m, m.shortID, cfg)
+	p := newFolder(m, m.shortID, cfg)
+	//p := newMasterFolder(m, m.shortID, cfg)
 	m.folderRunners[folder] = p
 	m.fmut.Unlock()
 
-	if len(cfg.Versioning.Type) > 0 {
+	if len(cfg.Versioning.Type) > 0 && !cfg.Master {
 		factory, ok := versioner.Factories[cfg.Versioning.Type]
 		if !ok {
 			l.Fatalf("Requested versioning type %q that does not exist", cfg.Versioning.Type)
@@ -185,7 +184,7 @@ func (m *Model) StartFolderDefault(folder string) {
 
 	m.Add(p)
 
-	l.Okln("Ready to synchronize", folder, "(read-write)")
+	l.Okln("Ready to synchronize", folder)
 }
 
 func (m *Model) warnAboutOverwritingProtectedFiles(folder string) {
@@ -214,29 +213,6 @@ func (m *Model) warnAboutOverwritingProtectedFiles(folder string) {
 	if len(filesAtRisk) > 0 {
 		l.Warnln("Some protected files may be overwritten and cause issues. See http://docs.syncthing.net/users/config.html#syncing-configuration-files for more information. The at risk files are:", strings.Join(filesAtRisk, ", "))
 	}
-}
-
-// StartFolderMaster starts read only processing on the current model. When in
-// read only mode the model will announce files to the cluster but not pull in
-// any external changes.
-func (m *Model) StartFolderMaster(folder string) {
-	m.fmut.Lock()
-	cfg, ok := m.folderCfgs[folder]
-	if !ok {
-		panic("cannot start nonexistent folder " + folder)
-	}
-
-	_, ok = m.folderRunners[folder]
-	if ok {
-		panic("cannot start already running folder " + folder)
-	}
-	s := newMasterFolder(m, m.shortID, cfg)
-	m.folderRunners[folder] = s
-	m.fmut.Unlock()
-
-	m.Add(s)
-
-	l.Okln("Ready to synchronize", folder, "(read only; no external updates accepted)")
 }
 
 type ConnectionInfo struct {
@@ -1776,11 +1752,7 @@ func (m *Model) CommitConfiguration(from, to config.Configuration) bool {
 			// A folder was added.
 			l.Debugln(m, "adding folder", folderID)
 			m.AddFolder(cfg)
-			if cfg.Master {
-				m.StartFolderMaster(folderID)
-			} else {
-				m.StartFolderDefault(folderID)
-			}
+			m.StartFolder(folderID)
 
 			// Drop connections to all devices that can now share the new
 			// folder.
